571: v 2.2: add client.py
